Use GraphQL client instead of custom model and IO (#9)
This uses the GraphQL client defined in the `cryoet-data-portal` package
related to this project. The main motivations for doing this are as
follows.

1. Reduce the code to maintain here.
2. Simplify keeping up with structural/schematic changes to the data
portal.
3. Exercise the code in `cryoet-data-portal` and improve it.

Listing the datasets and tomograms seems a little slower than using
s3fs/fsspec, but the benefits are worth the cost.

This removes support for the local filesystem as a storage backend for
the portal data because obviously the client does not support that. The
code to support different URLs for the GraphQL entry point are kept, but
we make the line-edit read-only in the GUI because only the default URL
works and others cause problems.

This also removes mocks from some of the tests because mocking the
client is trickier than mocking our own classes/functions. Tests take
longer, but we adjusted the timeouts to account for that. That's good
enough as a short term solution, but ultimately we need mocks/fakes to
avoid long running tests.