Add a check-box to control if existing layers are cleared (#15)
This adds a checkbox to the open panel that controls if the existing
layers are cleared before opening the next set of tomogram +
annotations. If it is checked, the layers are cleared. This takes effect
if the tomogram is opened by clicking on a tree item or by clicking the
open button. By default it is checked, so existing layers are cleared,
as the original purpose of this plugin was for browsing.

The was added to accommodate more complex workflows beyond simple
browsing (e.g. comparing point annotations with a automatic dense
segmentation from elsewhere).