Support dense segmentation and oriented-point annotations (#21)
This adds support for segmentation-mask and oriented-points annotations
types.

![Screenshot 2024-05-15 at 16 42
33](https://github.com/chanzuckerberg/napari-cryoet-data-portal/assets/2608297/dfeee38d-5d58-41c3-ba78-9f87afd589a0)

Segmentation masks are supported by loading the corresponding OME-Zarrs
in a similar way to the tomogram (i.e. according to resolution).

Oriented points are currently read and displayed as regular points. We
can follow up with another PR to augment that with the orientation
information (e.g. with a vectors layer).

As some datasets contain tomograms with lots of different annotations
(e.g. 10000), performance can suffer both while waiting for data to
arrive and when interacting with napari once it has.

This also fixes an alignment issue caused by the highest resolution
tomogram volume having a non-unit volume. I originally separated that
into #31, but there was some overlap with the changes here, so I brought
it into this PR.

Closes #22 and #23