Only rescale points for older napari versions (#33)
napari v0.5 includes a fix to handle expected combinations of point size
and layer scale. Previously, we worked around that issue by rescaling
the point sizes ourselves. After this PR, we only do that for older
versions of napari. The check is cached to avoid spamming the logs.

Alternatively, we could require `napari>=0.5` in `setup.cfg`. I decided
to support both versions for now, since v0.5 is quite new and comes with
a few of its own new bugs.

This was tested manually by installing v0.4.19 and v0.5.2 of napari and
visually inspecting the resulting point sizes.