Increase timeouts for tests that open and load data (#34)
Some tests run against the production API. Some of these have recently
been failing by timing out when waiting to open/load tomograms and their
annotations. This is occurring on `main` and I'd like to have the tests
pass there before creating a new release to address a couple of
unrelated bug fixes.

This PR increases timeouts for from 30 to 60 seconds. This is a short
term fix that I'm hoping will be good enough.

If we want to support this plugin and its tests in the long term, we
should probably set up a fake API for the tests to run against. Or
provide a way to filter which tomograms/annotations are opened/loaded
and use that in the tests.